Revert "integrate Qt graphics toolkit in Qt interpreter object"
This reverts commit 5dcc746ce6522d502e37d51ddf1a6131d4796dd8.

The original commit aimed to integrate QtHandles into the Octave
GUI library. This makes it impossible to use the qt backend without
the full Octave GUI, thus removing the possibility of keeping
QtHandles as a separate module.

This commit reversal is the first attempt at making QtHandles fully
modular again.